removes unnecessary files and places things in better places
