--- backEnd/environment.yaml
@@ -1,10 +1,0 @@
-name: lms
-channels:
-  - defaults
-dependencies:
-  - pymongo
-  - flask
-  - waitress
-  - nltk
-  - numpy
-  - scikit-learn
